requirements.txt
fastapi
uvicorn
sqlalchemy>=1.4
aiomysql
google-generativeai